fix: fix styles on Setting screen
