+++ common/theme.ts
@@ -10,6 +10,7 @@ export const theme = createTheme({
       error: '#A10035',
       background: '#eeeeee',
       grey0: 'rgba(255,255,255,0.75)',
+      grey1: 'rgba(0,0,0,0.5)',
       white: 'white',
       black: 'black',
    },
@@ -21,6 +22,7 @@ export const theme = createTheme({
       error: '#A10035',
       background: '#3FA796',
       grey0: 'rgba(255,255,255,0.75)',
+      grey1: 'rgba(0,0,0,0.5)',
       white: 'white',
       black: 'black',
    },
